+++ cmd/zetaclientd/initconfig.go
@@ -98,6 +98,7 @@ func InitializeConfig(_ *cobra.Command, _ []string) error {
 	configData.KeyringBackend = config.KeyringBackend(initializeConfigOpts.KeyringBackend)
 	configData.RelayerKeyPath = opts.RelayerKeyPath
 	configData.ComplianceConfig = sample.ComplianceConfig()
+	configData.FeatureFlags = sample.FeatureFlags()
 
 	// Save config file
 	return config.Save(&configData, globalOpts.ZetacoreHome)