+++ testutil/sample/zetaclient.go
@@ -47,3 +47,10 @@ func ComplianceConfig() config.ComplianceConfig {
 		},
 	}
 }
+
+// FeatureFlags returns a sample feature flags
+func FeatureFlags() config.FeatureFlags {
+	return config.FeatureFlags{
+		EnableMultipleCalls: true,
+	}
+}